requirements.txt
torch
torchvision
numpy
tqdm
scikit-learn
matplotlib
open_clip_torch
Pillow
pytorchvideo
fvcore
iopath
